+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `renderCallback()` method to apply customizations after rendering is complete. ([#79](https://github.com/walterra/d3-milestones/issues/79))
 - Added support for custom HTML ID attributes for milestone elements. ([#78](https://github.com/walterra/d3-milestones/issues/78))
+- Added WebP format to the list of supported image formats.
 
 ## [`v1.4.7`](https://github.com/walterra/d3-milestones/tree/v1.4.7)